+++ ft-session-replay/src/main/java/com/ft/sdk/sessionreplay/SessionReplayConstants.java
@@ -14,7 +14,7 @@ public class SessionReplayConstants {
 
     public static final String SESSION_REPLAY_BUS_MESSAGE_TYPE_KEY = "type";
     public static final String RUM_SESSION_RENEWED_BUS_MESSAGE = "rum_session_renewed";
-    public static final String RUM_KEEP_SESSION_BUS_MESSAGE_KEY = "keepSession";
+    public static final String RUM_KEEP_SESSION_BUS_COLLECT_TYPE_KEY = "collect_key";
     public static final String RUM_SESSION_ID_BUS_MESSAGE_KEY = "sessionId";
 
     public static final double DECREASE_PERCENT = 0.90;